tests(designer): update unit tests for some components (#269)
* config(designer): update test config files to support styles

This also changes the testMach location to run tests in the whole
application and not just the conditions file.

* refactor(designer): move componentListSelect files to own folder

* chore(designer): update componentListSelect file and file paths

Don't worry this was done automatically :)

* chore(designer): replace hapi test imports with testing library

This commit also removes a lot of async without await keywords.
There is a chance this commit will need to be reverted because
the asyncs were there for a specific reason but, oh well :)

* chore(designer): put back missing argument for save()

It looks like this argument is needed to update the component on the
designer, but it looks like the type definition doesn't expect an arg
so I get an error on my tslint. I'll have to look into this a bit more

* test(designer): finish updating tests for componentListSelect

This change componently removes any hapi helpers and enzyme. I haven't
yet removed Sinon because I don't know if it's necessary. I'll need to
look into it a bit more.

* docs(designer): add testing info to documentation

This also updates the eslintrc file to ignore markdown files and itself.
I don't know how commit were working before  🤷‍♂️

* test(designer): update tests for ComponentCreateList

* chore(designer): remove test limits and unused comment

* chore(designer): rename test files from .test to .jest

This is done because hapi test files have the .test extention and to
prevent confusion it makes sense to have a different extention for jest
or testing library files.

* chore(designer): removed sinon and stray console.log from tests

Yay no more sinon 🎉

Co-authored-by: Richard Oliver Bray <richardoliverbray@Richards-MBP.lan>