Allow save per page config on rehydration (#29)
In response to an Issue raised on: communitiesuk/funding-service-design-frontend#99

Problem definition
FSD frontend COPYs the forms jsons onto the form runner container as part of the DOCKER file workflow, called in this action.
However the form jsons references the application store as an output (for safe per page functionality) .
We need to load in the correct application store host ({{application_store_host}}=127.0.0.1:5001 locally) into the forms JSON callbackUrl (for the correct env), so when they are copied into the runner, it PUTs back to the correct application store endpoint on each page save.

Solution
The solution was to add a check for 'savePerPageWebhook == True' instead of a savePerPage URL. If set to True we use the callbackURL provided in the rehydration payload (if available). For our use case, the savePerPage Url and callbackUrl are the same: '{{application_store_host}}/applications/forms'.