FS-3664: read-only summary flag on the summary page
This introduces a new flag `is_read_only_summary` that can be passed in
the existing `metadata` payload when instantiating a new form. This
instructs the summary page to not show "Change" links on the questions
summary list and provides a direct back link.

Currently this fixes the back link and page title text for the summary
page according to if the flag is set but it would be preferable for
these values to be set using options bassed in my the service requesting
a new form rather than fixed in the form logic -- if there are existing
`options` payloads for customising this we should move to use those.

I'm applying the change to the existing custom form builder as well as
the newer form builder adpater to not be blocked on deployments as the
new patterns gains maturity and is released in different environments.

Note it doesn't look like individual flags like this are curretly
covered by tests even though they can change the behaviour of the
server, I'll look to do this in lining up with the methodology used by
the newer form builder adapter.